refactor: Move fe to more descriptive dir
