+++ client/fe/webpack/moduleRules.js
@@ -8,7 +8,7 @@
  */
 
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const config = require('../config').config
+const config = require('../config/config').config
 const resolveDir = require('./resolveDir').resolveDir
 
 /**
