+++ fe
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('./client/setup/fe')
+require('./client/fe/index')
